Break Exiter out into separate file
